url_shortener/url_shortener/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==3.2.25
django-cors-headers==4.4.0
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pytz==2024.2
redis==5.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
